refactor(tools): organize golden and calibration commands
