+++ docs/specs/razavi-visual-contract.md
@@ -40,7 +40,7 @@ PDF extraction, Symbol generation, and raster comparison are separate tools:
 ```text
 source PDF -> tools/pdf-vector-extract -> pinned vector evidence + PNG witness
 pinned vector evidence -> family generator -> Symbol DSL
-PNG witness + rendered Symbol -> scripts/razavi-fidelity-diff.mjs -> report
+PNG witness + rendered Symbol -> tools/calibration/razavi/fidelity-diff.mjs -> report
 ```
 
 The PDF extractor must not import the fidelity implementation, and the
@@ -256,7 +256,7 @@ pnpm --filter @icm/model build
 pnpm --filter @icm/derived build
 pnpm --filter @icm/render-svg build
 pnpm --filter @icm/exporters build
-node scripts/razavi-fidelity-diff.mjs <target>
+node tools/calibration/razavi/fidelity-diff.mjs <target>
 ```
 
 The PDF extraction command and dependencies are documented in
