+++ package.json
@@ -15,11 +15,11 @@
     "build": "pnpm -r --if-present build",
     "release:package": "pnpm build && node scripts/package-release.mjs",
     "performance:check": "pnpm build && node scripts/performance-baseline.mjs",
-    "export:phase7": "pnpm build && node scripts/phase-7-export-golden.mjs",
-    "export:phase7:check": "pnpm build && node scripts/phase-7-export-golden.mjs --check",
+    "export:golden": "pnpm build && node scripts/export-golden.mjs",
+    "export:golden:check": "pnpm build && node scripts/export-golden.mjs --check",
     "pwa:icons": "pnpm --filter @icm/exporters build && node scripts/pwa-icons.mjs",
     "pwa:icons:check": "pnpm --filter @icm/exporters build && node scripts/pwa-icons.mjs --check",
-    "release:verify:built": "node scripts/performance-baseline.mjs && node scripts/phase-7-export-golden.mjs --check && node scripts/pwa-icons.mjs --check && node scripts/editor-production-smoke.mjs --check && node scripts/package-release.mjs && node scripts/release-smoke.mjs",
+    "release:verify:built": "node scripts/performance-baseline.mjs && node scripts/export-golden.mjs --check && node scripts/pwa-icons.mjs --check && node scripts/editor-production-smoke.mjs --check && node scripts/package-release.mjs && node scripts/release-smoke.mjs",
     "release:verify": "pnpm build && pnpm release:verify:built",
     "agent-api:artifacts": "pnpm --filter @icm/agent-adapter... build && node scripts/agent-api-artifacts.mjs",
     "agent-api:artifacts:check": "pnpm --filter @icm/agent-adapter... build && node scripts/agent-api-artifacts.mjs --check",
@@ -45,8 +45,8 @@
     "symbols:razavi-opamp:check": "node scripts/generate-razavi-opamp-asset.mjs --check",
     "symbols:razavi-common": "node scripts/generate-razavi-common-assets.mjs",
     "symbols:razavi-common:check": "node scripts/generate-razavi-common-assets.mjs --check",
-    "visual:phase5": "pnpm build && node scripts/phase-5-golden.mjs",
-    "visual:phase5:check": "pnpm build && node scripts/phase-5-golden.mjs --check",
+    "visual:golden": "pnpm build && node scripts/visual-golden.mjs",
+    "visual:golden:check": "pnpm build && node scripts/visual-golden.mjs --check",
     "test": "vitest run",
     "test:local": "vitest run --maxWorkers=2",
     "test:e2e": "playwright test",
